feat: extract reflow and config out of editor.rs
